requirements.txt
# Flask Core
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7

# Security
Flask-Security-Too==5.4.3
bcrypt==4.2.1
cryptography>=41.0.0
python-dotenv==1.0.1
pyotp==2.9.0
qrcode==8.0
setuptools<81  # Pin to avoid pkg_resources deprecation warning

# Internationalization
Flask-Babel==4.0.0
Babel==2.16.0

# Database
SQLAlchemy==2.0.36

# Barcode Generation
python-barcode==0.15.1
Pillow==11.0.0

# Utilities
email-validator==2.2.0
WTForms==3.2.1

# Production Server
gunicorn==23.0.0

# Rate Limiting
Flask-Limiter==3.8.0
